Replace KJ_STACK_ARRAY with kj::SboArray/Builder
We'd like to be able to use KJ_STACK_ARRAY in coroutines, but KJ_STACK_ARRAY uses variable-length arrays on compilers which support them (that is: GCC). So, we'd like to drop support for VLAs from KJ_STACK_ARRAY.

There's not really any need for a macro anymore, if we don't support VLAs, however. So instead, we're replacing KJ_STACK_ARRAY entirely with the new SboArray and SboArrayBuilder class templates.

It's worth noting that at present, we don't actually support GCC, so we have no target which even uses VLAs anymore. But, we'd like to support GCC again in the future, once its coroutine support is fixed.