Refactor HeapArrayDisposer for reuse
HeapArrayDisposer contains a lot of fiddly logic to safely and optimally construct
and destroy array elements in the presence of trivial or noexcept constructors, as well as support for the ArrayDisposer use case. This refactor prepares the logic for later re-use in SboArrayDisposer, an ArrayDisposer which supports small-buffer optimization (SBO).